#8269. Replication, add more builders, updaters and first unit test.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@12458 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ frameworks/common-model/src/main/java/ish/oncourse/model/College.java
@@ -9,9 +9,9 @@ import java.util.Set;
 
 public class College extends _College {
 
-	public Integer getId() {
+	public Long getId() {
 		return (getObjectId() != null && !getObjectId().isTemporary()) ? ((Number) getObjectId()
-				.getIdSnapshot().get(ID_PK_COLUMN)).intValue() : null;
+				.getIdSnapshot().get(ID_PK_COLUMN)).longValue() : null;
 	}
 
 	public Set<String> getCollegeSiteStates() {
